separei as funções e estou criando os detalhes dos pokemons
