+++ detalhes.html
@@ -18,7 +18,10 @@
         <a href="./index.html">Home</a>
       </div>
     </header>
+    <div class="container">
+      <ul data="pokedex" class="pokedex"></ul>
+    </div>
 
-    <script src="./main.js"></script>
+    <script type="module" src="./detalhe.js"></script>
   </body>
 </html>
