+++ getPokemonUrl.js
@@ -1,0 +1,2 @@
+
+export const getPokemonUrl = id => `https://pokeapi.co/api/v2/pokemon/${id}`
