+++ index.html
@@ -20,6 +20,6 @@
         <ul data="pokedex" class="pokedex"></ul>
       </a>
     </div>
-    <script src="./main.js"></script>
+    <script type="module" src="./main.js"></script>
   </body>
 </html>
